add Stargaze blockchain (#31)
This PR adds Stargaze a Cosmos SDK based blockchain. [Cosmos SDK](cosmos.network) chains usually have the same APIs to retrieve the list of validators and voting power and I have abstracted the work made for cosmos hub chain to just point to the specific chain endpoint and name.

### About Stargaze
Stargaze is "The Interchain" NFT marketplace being the first interoperable layer-1 chain for NFTs. Stargaze is the app specific chain for NFTs in the Cosmos.

https://stargaze.zone
https://twitter.com/StargazeZone