remove ethereum PoS as the calculation is incorrect (#52)
Removes ethereum PoS as the calculation is incorrect.

NC calculation for ethereum is somewhat not straightforward. Eth staking
projects (like lido, stakewise) don't control the validators. The
operators running the validators control them and unfortunately we don't
have appropriate data source to measure this.

I believe EF is working on a data collections grant so that we have more
tools and data sources.