Add Cardano chain (#43)
**Description**: added Cardano chain.

**Ticket(s)**: #1 (closed but should be the actual one) and #16.

**Contribution Details**:
I used the [Cexplorer APIs](https://cexplorer.io/devs) to get the list
of pools and staked amount.

From Cexplorer, the Nakamoto coefficient is [something like
36](https://cexplorer.io/nakamoto). In the PR, anyway, the value you
will get is expected to be higher (the last time I tried was 112) as the
single pools are used instead of the
[groups](https://cexplorer.io/groups); that is what they used (confirmed
by an admin on Discord).